refactor(askiris): keep the client-facing type in the import-free module
LensLinkIndex is the type a client component holds, but it lived in the module
that reaches for the catalogue. `import type` is erased, so nothing shipped —
but one edit turning it into a value import would have pulled 477KB of JSON
into the AskIris bundle, with no error to catch it. lens-ref has no runtime
imports at all, which is where a type the client names belongs.
